Make {Add,Remove}FsmEdit instance methods of SilksongHost
