+++ ItemChanger.Silksong/SilksongHost.cs
@@ -13,6 +13,8 @@ namespace ItemChanger.Silksong
 {
     public class SilksongHost : ItemChangerHost
     {
+        public static SilksongHost Instance => (SilksongHost)ItemChangerHost.Singleton;
+
         internal SilksongHost() 
         {
             MessageUtil.Setup();
@@ -41,15 +43,32 @@ namespace ItemChanger.Silksong
         private LifecycleEvents.Invoker? lifecycleInvoker;
         private GameEvents.Invoker? gameInvoker;
 
+        public const string Wildcard = "*";
+
+        /// Registers a delegate to run whenever a FSM matching the given
+        /// (scene name, object name, FSM name) tuple is loaded.
+        /// The scene and object names can be Wildcard ("*") instead to match any scene or any
+        /// object, respectively.
+        public void AddFsmEdit(FsmId id, Action<PlayMakerFSM> edit)
+        {
+            fsmEdits[id] = fsmEdits.GetValueOrDefault(id) + edit;
+        }
+
+        public void RemoveFsmEdit(FsmId id, Action<PlayMakerFSM> edit)
+        {
+            fsmEdits[id] -= edit;
+        }
+
+        private Dictionary<FsmId, Action<PlayMakerFSM>?> fsmEdits = [];
+
         protected override void PrepareEvents(LifecycleEvents.Invoker lifecycleInvoker, GameEvents.Invoker gameInvoker)
         {
             this.lifecycleInvoker = lifecycleInvoker;
             this.gameInvoker = gameInvoker;
 
-            Type gmPatches = typeof(GameManagerPatches);
-            Harmony harmony = new(gmPatches.FullName);
-            harmony.PatchAll(gmPatches);
-            SilksongEvents.Hook();
+            Type patches = typeof(Patches);
+            Harmony harmony = new(patches.FullName);
+            harmony.PatchAll(patches);
             UnityEngine.SceneManagement.SceneManager.activeSceneChanged += OnActiveSceneChanged;
         }
 
@@ -58,8 +77,12 @@ namespace ItemChanger.Silksong
             this.lifecycleInvoker = null;
             this.gameInvoker = null;
 
-            Harmony.UnpatchID(typeof(GameManagerPatches).FullName);
-            SilksongEvents.Unhook();
+            foreach (var id in fsmEdits.Keys)
+            {
+                Logger.LogWarn($"FSM edit not cleaned up for {id.FsmName} in object {id.ObjectName} in scene {id.SceneName}");
+            }
+            fsmEdits.Clear();
+            Harmony.UnpatchID(typeof(Patches).FullName);
             UnityEngine.SceneManagement.SceneManager.activeSceneChanged -= OnActiveSceneChanged;
             MessageUtil.Clear();
         }
@@ -76,7 +99,7 @@ namespace ItemChanger.Silksong
         }
 
         [HarmonyPatch]
-        private static class GameManagerPatches
+        private static class Patches
         {
             [HarmonyPatch(typeof(GameManager), nameof(GameManager.StartNewGame))]
             [HarmonyPrefix]
@@ -138,6 +161,33 @@ namespace ItemChanger.Silksong
 
                 Host.gameInvoker?.NotifyBeforeNextSceneLoaded(new Events.Args.BeforeSceneLoadedEventArgs(targetScene)); // TODO: add gate info
                 // TODO: transition overrides
+            }
+
+            [HarmonyPatch(typeof(PlayMakerFSM), nameof(PlayMakerFSM.Start))]
+            [HarmonyPrefix]
+            private static void Prefix(PlayMakerFSM __instance)
+            {
+                var fsm = __instance;
+                var sceneName = fsm.gameObject.scene.name;
+                var objectName = fsm.gameObject.name;
+                var fsmName = fsm.FsmName;
+                List<FsmId> matchingIds = [
+                    new(sceneName, objectName, fsmName),
+                    new(Wildcard, objectName, fsmName),
+                    new(sceneName, Wildcard, fsmName),
+                    new(Wildcard, Wildcard, fsmName)
+                ];
+                try
+                {
+                    foreach (var id in matchingIds)
+                    {
+                        Instance.fsmEdits.GetValueOrDefault(id)?.Invoke(fsm);
+                    }
+                }
+                catch (Exception err)
+                {
+                    Instance.Logger.LogError($"Error applying FSM edit to FSM {fsmName} in object {objectName} in scene {sceneName}: {err}");
+                }
             }
         }
     }
